requirements.txt
# AndroSleuth - APK Forensic Analysis Tool
# Core dependencies

# APK Analysis
androguard>=3.4.0
apkid>=2.1.5

# Disassembly & Decompilation
capstone>=5.0.1
keystone-engine>=0.9.2
r2pipe>=1.8.0

# Binary Emulation
unicorn>=2.0.1

# Dynamic Instrumentation
frida>=16.1.10
frida-tools>=12.3.0

# Malware Detection
yara-python>=4.3.1

# Cryptography Analysis
pycryptodome>=3.19.0

# Network Analysis
mitmproxy>=10.1.5

# Utilities
colorama>=0.4.6
rich>=13.7.0
termcolor>=2.4.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.1.4
numpy>=1.26.2

# Visualization & Reporting
matplotlib>=3.8.2
jinja2>=3.1.2
plotly>=5.18.0

# Configuration
pyyaml>=6.0.1

# Database
sqlalchemy>=2.0.23

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
